requirements.txt
asyncio-nats-client==0.11.4
python-nmap==0.6.4